# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SIC API
  version: 0.1.0
  description: Implementation of accounting system based in hordak and following SIC
    (Andrisani) definitions.
  contact:
    name: gastong256
paths:
...
  /api/v1/companies/{company_id}/reports/journal-book/:
    get:
      operationId: get_journal_book
      description: Chronological list of all posted journal entries for the company
        within the given date range, including all lines and grand totals.
      summary: Libro Diario
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: date_from
        schema:
          type: string
        description: Start date (YYYY-MM-DD). Defaults to the earliest entry date.
      - in: query
        name: date_to
        schema:
          type: string
        description: End date (YYYY-MM-DD). Defaults to today.
      tags:
      - reports
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: Libro Diario report
        '400':
          description: Invalid query parameters
        '409':
          description: Company must be opened before reports are available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{company_id}/reports/journal-book.xlsx:
    get:
      operationId: export_journal_book_xlsx
      summary: Export Libro Diario (XLSX)
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: date_from
        schema:
          type: string
        description: Start date (YYYY-MM-DD). Defaults to the earliest entry date.
      - in: query
        name: date_to
        schema:
          type: string
        description: End date (YYYY-MM-DD). Defaults to today.
      tags:
      - reports
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: XLSX file
        '400':
          description: Invalid query parameters
        '409':
          description: Company must be opened before reports are available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{company_id}/reports/ledger/:
    get:
      operationId: get_ledger
      description: One card per level-3 account belonging to the company, showing
        every movement in the period chronologically with a running balance. Optionally
        filtered to a single account via account_id.
      summary: Libro Mayor
      parameters:
      - in: query
        name: account_id
        schema:
          type: integer
        description: Filter to a single level-3 account by its ID.
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: date_from
        schema:
          type: string
        description: Start date (YYYY-MM-DD). Defaults to the earliest entry date.
      - in: query
        name: date_to
        schema:
          type: string
        description: End date (YYYY-MM-DD). Defaults to today.
      - in: query
        name: include
        schema:
          type: string
        description: 'Optional extra sections. Supported values: account_options.'
      tags:
      - reports
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: Libro Mayor report
        '400':
          description: Invalid query parameters or account_id
        '409':
          description: Company must be opened before reports are available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{company_id}/reports/ledger.xlsx:
    get:
      operationId: export_ledger_xlsx
      summary: Export Libro Mayor (XLSX)
      parameters:
      - in: query
        name: account_id
        schema:
          type: integer
        description: Filter to a single level-3 account by its ID.
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: date_from
        schema:
          type: string
        description: Start date (YYYY-MM-DD). Defaults to the earliest entry date.
      - in: query
        name: date_to
        schema:
          type: string
        description: End date (YYYY-MM-DD). Defaults to today.
      tags:
      - reports
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: XLSX file
        '400':
          description: Invalid query parameters or account_id
        '409':
          description: Company must be opened before reports are available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{company_id}/reports/trial-balance/:
    get:
      operationId: get_trial_balance
      description: 'Two-level table: level-2 (colectiva) subtotal rows containing
        level-3 (subcuenta) rows, with total_debit, total_credit, and balance columns.
        Only accounts with movements in the period are included.'
      summary: Balance de Comprobación
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: date_from
        schema:
          type: string
        description: Start date (YYYY-MM-DD). Defaults to the earliest entry date.
      - in: query
        name: date_to
        schema:
          type: string
        description: End date (YYYY-MM-DD). Defaults to today.
      tags:
      - reports
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: Balance de Comprobación report
        '400':
          description: Invalid query parameters
        '409':
          description: Company must be opened before reports are available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{company_id}/reports/trial-balance.xlsx:
    get:
      operationId: export_trial_balance_xlsx
      summary: Export Balance de Comprobación (XLSX)
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: date_from
        schema:
          type: string
        description: Start date (YYYY-MM-DD). Defaults to the earliest entry date.
      - in: query
        name: date_to
        schema:
          type: string
        description: End date (YYYY-MM-DD). Defaults to today.
      tags:
      - reports
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: XLSX file
        '400':
          description: Invalid query parameters
        '409':
          description: Company must be opened before reports are available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{id}/:
    get:
      operationId: get_company
      summary: Retrieve a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
    put:
      operationId: update_company
      summary: Update a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyWrite'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '409':
          description: Read-only demo company
    patch:
      operationId: partial_update_company
      summary: Partially update a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCompanyWrite'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '409':
          description: Read-only demo company
    delete:
      operationId: delete_company
      summary: Delete a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No content
        '403':
          description: Permission denied
        '409':
          description: Company has protected accounting records
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
          description: Optional business description for a more realistic simulation.
        tax_id:
          type: string
          description: Simulated CUIT (optional).
          maxLength: 20
        owner_username:
          type: string
          readOnly: true
        account_count:
          type: integer
          readOnly: true
        books_closed_until:
          type: string
          format: date
          readOnly: true
          nullable: true
          description: Entries on or before this date are locked.
        is_demo:
          type: boolean
          readOnly: true
          description: Marks a seeded demo company shared across the system.
        is_read_only:
          type: boolean
          readOnly: true
          description: If true, accounting writes and metadata updates are blocked.
        has_opening_entry:
          type: boolean
          readOnly: true
        accounting_ready:
          type: boolean
          readOnly: true
        opening_entry_id:
          type: integer
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - account_count
      - accounting_ready
      - books_closed_until
      - created_at
      - has_opening_entry
      - id
      - is_demo
      - is_read_only
      - name
      - opening_entry_id
      - owner_username
      - updated_at
...
    CompanyWrite:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
      required:
      - name
...
    PatchedCompanyWrite:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT